organize for deployment
Former-commit-id: 0e9ac9b50c030d927c2b6f2e5f347b3444321903
Former-commit-id: 41dd97b25e9d7d025631963332f450bf86ede4fb

+++ database/migrations/2015_06_24_192650_create_episodes_table.php
@@ -26,7 +26,7 @@ class CreateEpisodesTable extends Migration
             $table->float('rating');
             $table->integer('votes')->unsigned();
             $table->integer('visits')->unsigned()->default('0');
-            $table->integer('order')->unsigned()->nullable();
+            $table->float('order')->unsigned()->nullable();
             $table->datetime('coming_date')->nullable();
             $table->timestamps();
             $table->softDeletes();
